requirements.txt
flask==2.3.3
python-dotenv==1.0.0
openai-agents
requests==2.31.0
anthropic==0.8.1
selenium==4.15.2
pillow==10.1.0
webdriver-manager==4.0.1 